fix: destructured AboutMePage
